% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control function.R
\name{control_function}
\alias{control_function}
\title{method2: control function}
\usage{
control_function(
  data,
  time,
  dow,
  temperature,
  resid1,
  optimal.df1,
  optimal.df2,
  resid2,
  pbl,
  ws,
  deaths
)
}
\arguments{
\item{data}{Data includes temperature, daily death toll, PM2.5 concentration,
time variable, boundary layer height and wind speed.}

\item{time}{Enter the time in a Date format, eg. 2016-01-01.}

\item{dow}{Enter the day of week variable in a factor format, eg. Monday.}

\item{temperature}{Enter the temperature in a numeric format, eg. 12.}

\item{resid1}{Separation function gets result.You can see the details in Separation.R and README.}

\item{optimal.df1}{The degrees of freedom of natural spline function with time variable
for predicting the deaths in a numeric format, eg. 15.}

\item{optimal.df2}{The degrees of freedom of natural spline function with temperature
for predicting the deaths in a numeric format, eg. 15.}

\item{resid2}{The residual resid1 and the instrument variables are obtained for support vector regression.}

\item{pbl}{Enter the height of the planetary boundary layer in a numeric format, eg. 302.}

\item{ws}{Enter the wind speed in a numeric format, eg. 0.8.}

\item{deaths}{The daily death toll in a numeric format, eg. 40.}
}
\value{
the health effect of pm2.5 and its 95\% CI by bootstrapping of time series.
}
\description{
method2: control function
}
