% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separation.R
\name{separation}
\alias{separation}
\title{separation function}
\usage{
separation(data, pm2.5, time, dow, temperature, optimal.df)
}
\arguments{
\item{data}{Data includes temperature, daily death toll, PM2.5 concentration, time variable, boundary layer height and wind speed.}

\item{pm2.5}{Enter the concentration of pm2.5 in a numeric format, eg. 150.}

\item{time}{Enter the time in a Date format, eg. 2016-01-01.}

\item{dow}{Enter the day of week variable in a factor format, eg. Monday.}

\item{temperature}{Enter the temperature in a numeric format, eg. 12.}

\item{optimal.df}{The degrees of freedom of natural spline function with time variable
for predicting the exposure pm2.5 in a numeric format, eg. 15.}
}
\value{
resid1-calculate the residual change of pm2.5 after controlling the confounding factors of
temperature and time trend.
}
\description{
separation function
}
\examples{

}
